+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## 0.1.6
+- Rock roll simulator
+
 ## 0.1.5
 - Format data, but seems to be at the cost of slower processing time
 - Fix monsterSkill hit ratio display